requirements.txt

attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.43.0
dash-core-components==0.48.0
dash-html-components==0.16.0
dash-renderer==0.24.0
dash-table==3.7.0
decorator==4.4.0
Flask==1.0.3
Flask-Compress==1.4.0
idna==2.8
iexfinance==0.4.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
lxml==4.3.3
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.3
pandas==0.24.2
pandas-datareader==0.7.0
plotly==3.9.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.1